+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Added
+* From Group Files: Ondersteuning voor een spinner (progress indicator) tijdens upload van bestand ([#2616](https://github.com/dso-toolkit/dso-toolkit/issues/2616))
+
 ## 🩹 62.20.1 - 26-04-2024
 
 ### Fixed